% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j15221_package.R
\docType{package}
\name{j15221_package}
\alias{j15221_package}
\alias{j15221}
\title{The 'j15221' package.}
\description{
Cpp and STAN functions for J15221 missing data analysis and
    hypothesis testing
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
